requirements.txt
# Backend dependencies
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-SocketIO==5.3.4
Flask-PyMongo==2.3.0
Flask-Limiter==3.3.1
Flask-Cors==4.0.0
pymongo==4.5.0
redis==4.6.0
python-socketio==5.8.0
gevent==23.9.1
gevent-websocket==0.10.1
gunicorn==21.2.0
APScheduler==3.10.1
Werkzeug==2.3.7
python-dotenv==1.0.0
requests==2.31.0
polyline==2.0.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0
xgboost==1.7.6
lifelines==0.27.4
matplotlib==3.7.2
seaborn==0.12.2

# Frontend dependencies
dash==2.13.0
dash-bootstrap-components==1.5.0
dash-leaflet==0.1.23
dash-extensions==1.0.4
plotly==5.16.1
Flask-Assets==2.0
webassets==2.0
libsass==0.22.0
cssmin==0.4.3
jsmin==3.0.1
pytz==2023.3
Pillow==10.0.0